requirements.txt
Adafruit-Blinka==8.58.1
adafruit-circuitpython-ahtx0==1.0.26
adafruit-circuitpython-am2320==1.2.24
adafruit-circuitpython-busdevice==5.2.12
adafruit-circuitpython-connectionmanager==3.1.4
adafruit-circuitpython-dht==4.0.8
adafruit-circuitpython-register==1.10.3
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.79.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
certifi==2025.4.26
charset-normalizer==3.4.2
colorzero==2.0
getrpimodel==0.1.26
gpio==1.0.0
gpiozero==2.0.1
grafana-client==4.0.0
idna==3.10
influxdb-client==1.38.0
lgpio==0.2.2.0
mh_z19==3.1.7
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
python-dotenv==1.0.1
requests==2.32.3
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing_extensions==4.13.2
urllib3==2.4.0
